Named route Post for User/Index/Admin

+++ routes/web.php
@@ -17,38 +17,38 @@ Route::get('/', function () {
 
 Route::group(['middleware' => 'user.auth'], function () {
     Route::group(['prefix' => 'post'], function () {
-        Route::get('/', 'PostController@index');
-        Route::get('/{id}', 'PostController@show');
-        Route::post('/comment/{id}', 'PostController@comment');
-        Route::get('/vote/{id}', 'PostController@vote');
+        Route::get('/', 'PostController@index')->name('post.index');
+        Route::get('/{id}', 'PostController@show')->name('post.show');
+        Route::post('/comment/{id}', 'PostController@comment')->name('post.comment');
+        Route::get('/vote/{id}', 'PostController@vote')->name('post.vote');
     });
 
     Route::group(['prefix' => 'user'], function () {
-        Route::get('/', 'UserController@listUser');
-        Route::get('/info', 'UserController@index');
-        Route::get('/follow/{id}', 'UserController@follow');
-        Route::get('/unfollow/{id}', 'UserController@unfollow');
+        Route::get('/', 'UserController@listUser')->name('user.list');
+        Route::get('/info', 'UserController@index')->name('user.index');
+        Route::get('/follow/{id}', 'UserController@follow')->name('user.follow');
+        Route::get('/unfollow/{id}', 'UserController@unfollow')->name('user.unfollow');
 
         Route::group(['prefix' => 'post', 'namespace' => 'User'], function () {
-            Route::get('/', 'PostController@index');
-            Route::get('/create', 'PostController@create')->middleware('user.block');
-            Route::get('/{id}', 'PostController@show');
-            Route::post('/create', 'PostController@store')->middleware('user.block');
-            Route::get('/{id}/edit', 'PostController@edit');
-            Route::delete('/{id}', 'PostController@destroy');
-            Route::put('/{id}', 'PostController@update');
+            Route::get('/', 'PostController@index')->name('user.post.index');
+            Route::get('/create', 'PostController@create')->name('user.post.create')->middleware('user.block');
+            Route::get('/{id}', 'PostController@show')->name('user.post.show');
+            Route::post('/create', 'PostController@store')->name('user.post.store')->middleware('user.block');
+            Route::get('/{id}/edit', 'PostController@edit')->name('user.post.edit');
+            Route::delete('/{id}', 'PostController@destroy')->name('user.post.delete');
+            Route::put('/{id}', 'PostController@update')->name('user.post.update');
         });
     });
 
     Route::group(['prefix' => 'category'], function () {
-       Route::get('/', 'CategoryController@index');
+        Route::get('/', 'CategoryController@index');
     });
 });
 
 Route::group(['prefix' => 'auth'], function () {
     Route::get('register', 'UserController@showRegister');
     Route::post('register', 'UserController@register');
-    Route::get('login', 'UserController@showLogin');
+    Route::get('login', 'UserController@showLogin')->name('auth.login');
     Route::post('login', 'UserController@login')->name('auth.login');
     Route::get('logout', 'UserController@logout');
 });
@@ -70,10 +70,10 @@ Route::group(['prefix' => 'admin', 'middleware' => 'admin.auth'], function () {
         });
 
         Route::group(['prefix' => 'post'], function () {
-            Route::get('/', 'PostController@all');
-            Route::get('/{id}', 'PostController@show');
-            Route::delete('/{id}','PostController@delete');
-            Route::get('/restore/{id}', 'PostController@restore');
+            Route::get('/', 'PostController@all')->name('admin.post');
+            Route::get('/{id}', 'PostController@show')->name('admin.post.show');
+            Route::delete('/{id}', 'PostController@delete')->name('admin.post.delete');
+            Route::get('/restore/{id}', 'PostController@restore')->name('admin.post.restore');
         });
     });
 });
